Add a dummy test
Why:

It's always useful to activate the tests, even for trivial methods. A dummy test is a good start.

This change addresses the need by:

This dummy test needs some modules to be mocked. What a pain! I'm using
inject-loader to be able to substitute 'pixi.js' by a mock. ESLint
complains about the code tied to webpack, and I cannot disagree. The
errors raised by ESLint are disabled with specific comments.

I'm not finished with this issue. I'm sure there's a better way to mock an
imported module.