Position of the teeth on the comb
Why:

The position of the teeth depends of the position of the comb. Adjustments
should be possible.

This change addresses the need by:

Create a teeth container that can be moved easily. Add some options to
adjust the position of the teeth container.

The size of the teeth depends of the available space left by the comb,
minus the adjustments.