Move the pins container in the comb
Why:

The pins have to be aligned with the teeth. If the pins are totally
separated from the teeth, it is quite difficult to align them.

This change addresses the need by:

The pins container has been moved to the comb. The comb is becoming the
graphic part of the music box, while the box itself is becoming the glue
between the graphics and the sequencer.