Change the skins
Change the skins of the box. Populate the <select> input with available boxes. set the Dummy box by default.

Because the sequencer and the synth are reloaded each time, the sound goes weird after changing the skins many times. The music should stop and the sequencer should be reset after changing the box.